소스 개선

+++ CryptokiMan/decrypt_dlg.cpp
@@ -862,7 +862,7 @@ void DecryptDlg::runFileDecrypt()
     int nPartSize = manApplet->settingsMgr()->fileReadSize();
     qint64 nReadSize = 0;
     int nLeft = 0;
-    int nOffset = 0;
+    qint64 nOffset = 0;
     int nPercent = 0;
 
     QString strSrcFile = mSrcFileText->text();
